Tests are devided into several modules
